SpartanUP: xilnvm: Initialize Efuse clock source control register
If XNVM_SET_EFUSE_CLK_FREQUENCY_FROM_RTCA is not set, set the frequency
clock source value provided by user in XNVM_EFUSE_CLK_CTRL_ADDR register.

Acked-by: Durga Challa <durga.challa@amd.com>

+++ lib/sw_services/xilnvm/src/spartanup/xnvm_efuse.h
@@ -422,7 +422,7 @@ typedef struct {
 #ifdef SPARTANUPLUSAES1
 	XNvm_EfuseBootModeDis *BootModeDis; /**< Pointer to the boot mode disable structure */
 #endif
-	u64 EfuseClkFreq; /**< eFuse clock frequency */
+	u32 EfuseClkFreq; /**< eFuse clock frequency */
 	u32 EfuseClkSrc; /**< eFuse clock source */
 } XNvm_EfuseData;
